+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.9.1] - 2022-07-31
+### Fixed
+- Dropped tiles not at the highest most point
+### Added
+- Support for thumb sizes (increase/decrease)
+
 ## [4.8.0] - 2022-05-15
 ### Changed
 - Manage sources (for indexing process)